chore: consolidate multiple files in tests/connect
