+++ tests/connect/test_config.py
